build: big update to build process, use esbuild with vite based site usage
