+++ .gitignore
@@ -22,3 +22,9 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+# TS compiled
+**/**/*.js
+!scripts/**/*.js
+**/**/*.js.map
+**/**/*.d.ts
