+++ package.json
@@ -1,12 +1,48 @@
 {
-  "name": "gcp-editor",
-  "private": true,
+  "name": "@hotosm/gcp-editor",
   "version": "0.0.0",
-  "type": "commonjs",
+  "description": "A web component for working with Ground Control Points.",
+  "homepage": "https://github.com/hotosm/gcp-editor#readme",
+  "author": {
+    "name": "HOTOSM",
+    "email": "sysadmin@hotosm.org"
+  },
+  "contributors": [
+    {
+      "name": "Sujit",
+      "url": "https://github.com/Sujit"
+    },
+    {
+      "name": "Sam Woodcock",
+      "url": "https://github.com/spwoodcock"
+    }
+  ],
+  "license": "AGPL-3.0-only",
+  "main": "./dist/gcp-editor.js",
+  "module": "./dist/gcp-editor.js",
+  "type": "module",
+  "types": "./dist/gcp-editor.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/gcp-editor.d.ts",
+      "import": "./dist/gcp-editor.js",
+      "default": "./dist/gcp-editor.js"
+    },
+    "./style.css": "./dist/gcp-editor.css"
+  },
+  "files": [
+    "dist"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/hotosm/gcp-editor.git"
+  },
   "scripts": {
-    "dev": "vite --host 0.0.0.0 --port 3000",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
+    "build:dts": "tsc --emitDeclarationOnly",
+    "build:js": "scripts/build.js",
+    "build": "pnpm run build:js",
+    "dev": "pnpm run build && pnpm run --dir site dev",
+    "preview": "pnpm run build && pnpm run --dir site preview"
   },
   "dependencies": {
     "@hotosm/ui": "^0.2.0-b6",
@@ -18,12 +54,19 @@
   },
   "devDependencies": {
     "autoprefixer": "^10.4.20",
+    "esbuild": "^0.24.0",
+    "esbuild-plugin-tailwindcss": "^1.2.1",
     "postcss": "^8.4.49",
     "tailwindcss": "^3.4.16",
     "typescript": "~5.6.2",
     "vite": "^5.4.10"
   },
   "peerDependencies": {
+    "@shoelace-style/shoelace": "^2.17.1",
     "ol": "^7.5.0"
+  },
+  "peerDependencies: comments": {
+    "ol": "PINNED TO MATCH @openlayers-elements/core",
+    "@shoelace-style/shoelace": "PINNED TO MATCH @hotosm/ui"
   }
 }
