+++ site/README.md
@@ -1,0 +1,27 @@
+# GCP Editor Example Site
+
+- Built to Github pages.
+- A demo of the GCP Editor component.
+
+## Install
+
+```bash
+npm install @hotosm/gcp-editor
+```
+
+## Usage
+
+```html
+<!doctype html>
+<html lang="en">
+  <head>
+    <script type="module">
+      import '@hotosm/gcp-editor';
+      import '@hotosm/gcp-editor/style.css';
+    </script>
+  </head>
+  <body>
+    <gcp-editor></gcp-editor>
+  </body>
+</html>
+```
