+++ site/index.html
@@ -5,25 +5,12 @@
     <link rel="icon" type="image/svg+xml" href="/hot-logo-icon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>GCP Editor</title>
-    <link rel="stylesheet" href="./src/index.css" />
-    <link rel="stylesheet" href="./src/styles/tailwind.css" id="tailwind-styles"/>
     <link href="https://fonts.cdnfonts.com/css/barlow-condensed" rel="stylesheet">
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined" id="material-symbols"/>
-    <!-- <script type="module" src="/src/gcp-editor.ts"></script> -->
-    <script type="module" src="./src/gcp-editor-wrapper.ts"></script>
-    <style>
-      *,
-     *::before,
-     *::after {
-       box-sizing: border-box;
-       padding: 0;
-       margin: 0;
-     }
-     .main-component-wrapper{
-       height:100vh;
-       width:100vw;
-     }
-   </style>
+    <script type="module">
+      import '@hotosm/gcp-editor';
+      import '@hotosm/gcp-editor/style.css';
+    </script>
   </head>
   <body>
     <div class="main-component-wrapper">
