+++ src/style.css
@@ -1,3 +1,7 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
 .image-with-outer-wrapper{
     width: 650px;
     height: 430px;
@@ -33,6 +37,18 @@
     overflow-y:auto;
    }
    
-   #image-uploading-content {
+  #image-uploading-content {
     height: 80vh;
+  }
+
+*,
+  *::before,
+  *::after {
+    box-sizing: border-box;
+    padding: 0;
+    margin: 0;
+  }
+  .main-component-wrapper{
+    height:100vh;
+    width:100vw;
   }
